Restructure: extract RaptorCache + RaptorCachePool + PooledCache into cache.rs
